requirements-poc.txt
mlflow==2.9.2
scikit-learn==1.3.2
xgboost==2.0.3
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1
cassandra-driver==3.29.0
feast==0.45.0
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-multipart==0.0.6
prometheus-client==0.19.0
onnx==1.15.0
onnxruntime==1.16.3
skl2onnx==1.16.0
requests==2.31.0
matplotlib==3.8.2
seaborn==0.13.1
tabulate==0.9.0
tqdm==4.66.1
pytest==9.0.2
